requirements.txt
PyQt6>=6.6.0
google-generativeai>=0.8.0
ffmpeg-python>=0.2.0
pydub>=0.25.1
opencv-python>=4.9.0
python-dotenv>=1.0.0
numpy>=1.24.0
Pillow>=10.0.0
